# swagger.yml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "www.virtual-wallet.io"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /accounts:

    post:
      tags:
      - "Accounts"
      summary: "Create a new account"
      description: ""
      operationId: "addAccount"
      consumes:
      - "application/json"
      - "application/xml"
      parameters:
        - in: "body"
          name: "user"
          schema:
            $ref: "#/definitions/User"
          required: true
      produces:
      - "application/xml"
      - "application/json"
      responses:
        405:
          description: "Invalid input"
        403:
          description: "User doesn't exist or Account exists"
      security:
      - wallet_auth:
        - "write:account"
        - "read:account"

    get:
      tags:
      - "Accounts"
      summary: "Get all accounts"
      description: ""
      operationId: "getAccount"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        200:
          description: OK
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Account"
        405:
          description: "Invalid input"
      security:
      - wallet_auth:
        - "write:account"
        - "read:account"
  /accounts/{accountId}:

    get:
      tags:
      - "Account"
      summary: "Get details of account"
      description: ""
      operationId: "getAccountDetails"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "accountId"
        in: "path"
        description: "Status values that need to be considered for filter"
        required: true
        type: "number"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Account"
        400:
          description: "Invalid status value"
      security:
      - wallet_auth:
        - "write:account"
        - "read:account"

    delete:
      tags:
      - "Account"
      summary: "Close account"
      description: ""
      operationId: "closeAccount"
      parameters:
        - name: "accountId"
          type: "number"
          in: "path"
          required: true
      responses:
        200:
          description: OK
        400:
          description: "Invalid account ID"
        405:
          description: "Invalid authorisation"
      security:
        - wallet_auth:
          - "write:account"
    
    post:
      tags:
      - "Account"
      summary: "Create a new transaction"
      description: ""
      operationId: "createAccountTransaction"
      consumes:
      - "application/Json"
      parameters:
        - name: "accountId"
          type: "number"
          in: "path"
          required: true
        - in: "body"
          name: "account"
          schema:
            $ref: "#/definitions/Account"
      responses:
        200:
          description: OK
        400:
          description: "Invalid account ID"
        405:
          description: "Invalid authorisation"
      security:
        - wallet_auth:
          - "write:transaction"
  /accounts/{accountId}/{transactionId}:

    get:
      tags:
        - "Account"
      summary: "Get transaction details"
      operationId: "getUserTransactionDetails"
      produces:
      - "application/json"
      parameters:
        - name: "accountId"
          in: "path"
          required: true
          type: "number"
        - name: "transactionId"
          in: "path"
          type: "string"
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Transaction"
      deprecated: false
      security:
        - wallet_auth:
          - "read:transaction"

    delete:
      tags:
        - "Account"
      summary: "Reverse transaction"
      parameters:
        - name: "accountId"
          in: "path"
          required: true
          type: "number"
        - name: "transactionId"
          in: "path"
          type: "string"
          required: true
      responses:
        200:
          description: OK
        402:
          description: "Invalid authentication"
      deprecated: false
      security:
        - wallet_auth:
          - "write:transaction"
...
securityDefinitions:
  wallet_auth:
    type: "oauth2"
    authorizationUrl: "https://www.virtual-wallet.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:user: "modify user"
      read:user: "read user"
      write:account: "modify account"
      read:account: "read your account"
      write:transaction: "modify transaction"
      read:transaction: "read transaction"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  User:
    type: "object"
    properties:
      userId:
        type: "integer"
        format: "int64"
      firstName:
        type: "string"
      surname:
        type: "string"
      email:
        type: "string"
        format: "email"
      phoneNumber:
        type: "string"
      password:
        type: "string"
      accounts:
        type: "array"
        format: "number"
    xml:
      name: "User"
  Transaction:
    type: "object"
    properties:
      transactionId:
        type: "string"
      fromAccount:
        type: "object"
        properties:
          accountNumber:
            type: "string"
          openingBalance:
            type: "integer"
            format: "int64"
      toAccount:
        type: "object"
        properties:
          accountNumber:
            type: "string"
          openingBalance:
            type: "integer"
            format: "int64"
      dateInitiation:
        type: "string"
        format: "date-time"
      dateSettlement:
        type: "string"
        format: "date-time"
      settled:
        type: "boolean"
        default: false
    xml:
      name: "Transaction"
  Account:
    type: "object"
    properties:
      accountNumber:
        type: "integer"
        format: "int64"
      balance:
        type: "integer"
        format: "int64"
      availableBalance:
        type: "integer"
        format: "int64"
      closed:
        type: "boolean"
        default: false
      transactions:
        type: "array"
        format: "string"
    xml:
      name: "Account"
...